Move a lot of files
