+++ telegram/__init__.py
@@ -18,23 +18,19 @@
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
 """A library that provides a Python interface to the Telegram Bot API"""
 
-from sys import version_info
-import sys
-import os
-
 from .base import TelegramObject
 from .user import User
 from .chat import Chat
 from .chatmember import ChatMember
-from .photosize import PhotoSize
-from .audio import Audio
-from .voice import Voice
-from .document import Document
-from .sticker import Sticker
-from .video import Video
-from .contact import Contact
-from .location import Location
-from .venue import Venue
+from .files.photosize import PhotoSize
+from .files.audio import Audio
+from .files.voice import Voice
+from .files.document import Document
+from .files.sticker import Sticker
+from .files.video import Video
+from .files.contact import Contact
+from .files.location import Location
+from .files.venue import Venue
 from .chataction import ChatAction
 from .userprofilephotos import UserProfilePhotos
 from .keyboardbutton import KeyboardButton
@@ -43,56 +39,56 @@ from .replykeyboardmarkup import ReplyKeyboardMarkup
 from .replykeyboardremove import ReplyKeyboardRemove, ReplyKeyboardHide
 from .forcereply import ForceReply
 from .error import TelegramError
-from .inputfile import InputFile
-from .file import File
+from .files.inputfile import InputFile
+from .files.file import File
 from .emoji import Emoji
 from .parsemode import ParseMode
 from .messageentity import MessageEntity
-from .animation import Animation
-from .game import Game
-from .shippingaddress import ShippingAddress
-from .orderinfo import OrderInfo
-from .successfulpayment import SuccessfulPayment
-from .invoice import Invoice
+from .games.animation import Animation
+from .games.game import Game
+from .payment.shippingaddress import ShippingAddress
+from .payment.orderinfo import OrderInfo
+from .payment.successfulpayment import SuccessfulPayment
+from .payment.invoice import Invoice
 from .message import Message
-from .inputmessagecontent import InputMessageContent
 from .callbackquery import CallbackQuery
 from .choseninlineresult import ChosenInlineResult
-from .inlinekeyboardbutton import InlineKeyboardButton
-from .inlinekeyboardmarkup import InlineKeyboardMarkup
-from .inlinequery import InlineQuery
-from .inlinequeryresult import InlineQueryResult
-from .inlinequeryresultarticle import InlineQueryResultArticle
-from .inlinequeryresultaudio import InlineQueryResultAudio
-from .inlinequeryresultcachedaudio import InlineQueryResultCachedAudio
-from .inlinequeryresultcacheddocument import InlineQueryResultCachedDocument
-from .inlinequeryresultcachedgif import InlineQueryResultCachedGif
-from .inlinequeryresultcachedmpeg4gif import InlineQueryResultCachedMpeg4Gif
-from .inlinequeryresultcachedphoto import InlineQueryResultCachedPhoto
-from .inlinequeryresultcachedsticker import InlineQueryResultCachedSticker
-from .inlinequeryresultcachedvideo import InlineQueryResultCachedVideo
-from .inlinequeryresultcachedvoice import InlineQueryResultCachedVoice
-from .inlinequeryresultcontact import InlineQueryResultContact
-from .inlinequeryresultdocument import InlineQueryResultDocument
-from .inlinequeryresultgif import InlineQueryResultGif
-from .inlinequeryresultlocation import InlineQueryResultLocation
-from .inlinequeryresultmpeg4gif import InlineQueryResultMpeg4Gif
-from .inlinequeryresultphoto import InlineQueryResultPhoto
-from .inlinequeryresultvenue import InlineQueryResultVenue
-from .inlinequeryresultvideo import InlineQueryResultVideo
-from .inlinequeryresultvoice import InlineQueryResultVoice
-from .inlinequeryresultgame import InlineQueryResultGame
-from .inputtextmessagecontent import InputTextMessageContent
-from .inputlocationmessagecontent import InputLocationMessageContent
-from .inputvenuemessagecontent import InputVenueMessageContent
-from .inputcontactmessagecontent import InputContactMessageContent
-from .labeledprice import LabeledPrice
-from .shippingoption import ShippingOption
-from .precheckoutquery import PreCheckoutQuery
-from .shippingquery import ShippingQuery
+from .inline.inlinekeyboardbutton import InlineKeyboardButton
+from .inline.inlinekeyboardmarkup import InlineKeyboardMarkup
+from .inline.inputmessagecontent import InputMessageContent
+from .inline.inlinequery import InlineQuery
+from .inline.inlinequeryresult import InlineQueryResult
+from .inline.inlinequeryresultarticle import InlineQueryResultArticle
+from .inline.inlinequeryresultaudio import InlineQueryResultAudio
+from .inline.inlinequeryresultcachedaudio import InlineQueryResultCachedAudio
+from .inline.inlinequeryresultcacheddocument import InlineQueryResultCachedDocument
+from .inline.inlinequeryresultcachedgif import InlineQueryResultCachedGif
+from .inline.inlinequeryresultcachedmpeg4gif import InlineQueryResultCachedMpeg4Gif
+from .inline.inlinequeryresultcachedphoto import InlineQueryResultCachedPhoto
+from .inline.inlinequeryresultcachedsticker import InlineQueryResultCachedSticker
+from .inline.inlinequeryresultcachedvideo import InlineQueryResultCachedVideo
+from .inline.inlinequeryresultcachedvoice import InlineQueryResultCachedVoice
+from .inline.inlinequeryresultcontact import InlineQueryResultContact
+from .inline.inlinequeryresultdocument import InlineQueryResultDocument
+from .inline.inlinequeryresultgif import InlineQueryResultGif
+from .inline.inlinequeryresultlocation import InlineQueryResultLocation
+from .inline.inlinequeryresultmpeg4gif import InlineQueryResultMpeg4Gif
+from .inline.inlinequeryresultphoto import InlineQueryResultPhoto
+from .inline.inlinequeryresultvenue import InlineQueryResultVenue
+from .inline.inlinequeryresultvideo import InlineQueryResultVideo
+from .inline.inlinequeryresultvoice import InlineQueryResultVoice
+from .inline.inlinequeryresultgame import InlineQueryResultGame
+from .inline.inputtextmessagecontent import InputTextMessageContent
+from .inline.inputlocationmessagecontent import InputLocationMessageContent
+from .inline.inputvenuemessagecontent import InputVenueMessageContent
+from .inline.inputcontactmessagecontent import InputContactMessageContent
+from .payment.labeledprice import LabeledPrice
+from .payment.shippingoption import ShippingOption
+from .payment.precheckoutquery import PreCheckoutQuery
+from .payment.shippingquery import ShippingQuery
 from .webhookinfo import WebhookInfo
-from .gamehighscore import GameHighScore
-from .videonote import VideoNote
+from .games.gamehighscore import GameHighScore
+from .files.videonote import VideoNote
 from .update import Update
 from .bot import Bot
 from .constants import (MAX_MESSAGE_LENGTH, MAX_CAPTION_LENGTH, SUPPORTED_WEBHOOK_PORTS,
