+++ telegram/files/photosize.py
