+++ telegram/files/sticker.py
