+++ telegram/inline/inlinequeryresultcacheddocument.py
