+++ telegram/inline/inlinequeryresultcachedgif.py
