+++ telegram/inline/inlinequeryresultcachedphoto.py
