+++ telegram/inline/inlinequeryresultcachedsticker.py
