+++ telegram/inline/inlinequeryresultcachedvoice.py
