+++ telegram/inline/inlinequeryresultdocument.py
