+++ telegram/inline/inlinequeryresultgame.py
