+++ telegram/inline/inlinequeryresultlocation.py
