+++ telegram/inline/inlinequeryresultphoto.py
