+++ telegram/inline/inlinequeryresultvoice.py
