+++ telegram/inline/inputlocationmessagecontent.py
