+++ telegram/inline/inputtextmessagecontent.py
