+++ telegram/payment/precheckoutquery.py
