+++ telegram/payment/shippingquery.py
